requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.0
boto3==1.34.113
mlflow==2.12.1
requests==2.32.3
python-dotenv==1.0.1 